// apps/web/src/types/module.ts
export type ModuleStatus = 'draft' | 'beta' | 'pilot' | 'published'

export type ModuleRating = 1 | 2 | 3 | 4 | 5

export type ReactionType =
    | 'like'
    | 'love'
    | 'curious'
    | 'confused'
    | 'insightful'

export type ModuleRequiredSubscription = 'free' | 'premium'

type ModuleMetadata = {
    generalInfo: {
        createdAt: Date
        updatedAt?: Date
        version: string // Format: '1.0.0'
        serialNumber: number // Unique identifier
        status: ModuleStatus
        authors: string[]
        contributors?: string[]
    }
    analyticsData?: {
        views: number
        completions: number
        averageCompletionTime?: number
        rating?: ModuleRating
        feedbackCount?: number
        dropoffRate?: number
    }
    accessControl: {
        isPublic: boolean
        allowedUsers?: string[] // User ids
        requiredSubscription?: ModuleRequiredSubscription
        accessLogs?: {
            userId: string
            accessedAt: Date
        }[]
    }
    contentMetadata: {
        title: string
        categories: string[] // Categories
        tags: string[] // Keywords
        prerequisites?: string[] // Modules ids
        shortDescription: string // 160 characters
        longDescription: string // 500 characters
        objectives: string[] // Learning outcomes of the module (e.g. 'Understand the basics of React')
        subjects: string[] // Subjects covered in the module (e.g. 'React', 'Redux')
        estimatedCompletionTime?: number // In seconds
        language: string // Hebrew, English, etc.
        betaVersionUrl?: string // URL to the beta version of the module. This is the link to the Google Docs Document that contains the module content.
    }
    engagement?: {
        likes: number // Number of users who liked the module
        favorites: number // Number of users who saved the module
        shares: number // Number of users who shared the module
        comments: {
            userId: string
            username: string
            timestamp: Date
            text: string
            likes?: number
            replies?: {
                userId: string
                username: string
                timestamp: Date
                text: string
                likes?: number
            }[]
        }[]
        discussions?: {
            topic: string
            participants: string[] // User ids
            messages: {
                sender: string
                timestamp: Date
                text: string
                attachments?: string[]
            }[]
        }[]
        reactions?: {
            userId: string
            type: ReactionType
            timestamp: Date
        }[]
        userProgressStats?: {
            totalEnrolled: number
            activeUsers: number
            averageCompletionTime?: number
            highestRatedUnit?: string // Unit id
        }
        leaderboard?: {
            userId: string
            username: string
            points: number // Points earned by the user
            completedModules: number // Number of modules completed by the user
            badges?: string[] // Badges ids
        }[]
    }
    evolution?: {
        featureRequests?: string[]
        improvementSuggestions?: string[]
        bugReports?: {
            issue: string
            reportedBy: string // User id
            status: 'open' | 'in_progress' | 'resolved'
            createdAt: Date
        }[]
    }
}

export type Module = {
    _id: string
    metadata: ModuleMetadata
    units?: Unit[]
}

type UnitMetaData = {
    generalInfo: {
        createdAt: Date
        updatedAt?: Date
        status: ModuleStatus
        authors: string[]
        contributors?: string[]
    }
    analyticsData: {
        views: number
        completions: number
        averageCompletionTime?: number
        rating?: ModuleRating
        feedbackCount?: number
        dropoffRate?: number
    }
    accessControl: {
        isPublic: boolean
        allowedUsers?: string[] // User ids
        requiredSubscription?: ModuleRequiredSubscription
        accessLogs?: {
            userId: string
            accessedAt: Date
        }[]
    }
    contentMetadata: {
        title: string
        categories: string[] // Categories
        tags: string[] // Keywords
        shortDescription: string // 160 characters
        longDescription: string // 500 characters
        objectives: string[] // Learning outcomes of the unit (e.g. 'Understand the basics of React')
        subjects: string[] // Subjects covered in the unit (e.g. 'React', 'Redux')
        estimatedCompletionTime?: number // In seconds
    }
    engagement: {
        likes: number // Number of users who liked the module
        favorites: number // Number of users who saved the module
        shares: number // Number of users who shared the module
        comments: {
            userId: string
            username: string
            timestamp: Date
            text: string
            likes?: number
            replies?: {
                userId: string
                username: string
                timestamp: Date
                text: string
                likes?: number
            }[]
        }[]
        discussions?: {
            topic: string
            participants: string[] // User ids
            messages: {
                sender: string
                timestamp: Date
                text: string
                attachments?: string[]
            }[]
        }[]
        reactions?: {
            userId: string
            type: ReactionType
            timestamp: Date
        }[]
        userProgressStats?: {
            totalEnrolled: number
            activeUsers: number
            averageCompletionTime?: number
        }
    }
    evolution: {
        featureRequests?: string[]
        improvementSuggestions?: string[]
        bugReports?: {
            issue: string
            reportedBy: string // User id
            status: 'open' | 'in_progress' | 'resolved'
            createdAt: Date
        }[]
    }
}

export type Unit = {
    _id: string
    metadata: UnitMetaData
    levels: Level[]
}
// ...
export type Level = {
    _id: string
    metadata: LevelMetaData
    steps: Step[]
}